v4/python/requirements-ai.txt
# Core (Required)
numpy>=1.24.0
cryptography>=3.4.7
aiohttp>=3.8.0
psutil>=5.9.0
pyyaml>=6.0
colorama>=0.4.6
tqdm>=4.65.0
blake3>=0.3.3

# TUI (Optional but recommended)
rich>=13.0.0

# AI/ML (Required for Elara)
torch>=2.0.0
transformers>=4.30.0
tiktoken>=0.5.0
sentencepiece>=0.1.99
protobuf>=3.20.0

# RAG/Vector DB (Required for Wolverine, HiveMind)
chromadb>=0.4.0
tiktoken>=0.5.0

# OCR (Required for Owl)
pytesseract>=0.3.10
Pillow>=10.0.0
pypdf>=3.0.0

# Audio (Required for Bat/Whistle)
pyaudio>=0.2.11
scipy>=1.10.0

# LLM Backends (At least one required)
ollama>=0.1.0

# Container (Optional, for Octopus)
aiodocker>=0.21.0

# Build Tools
pyinstaller>=5.0
